Use get_op_result_or_value for slice_sizes in gather, fix failing unit test
