+++ tripy/tripy/flat_ir/ops/concatenate.py
@@ -37,10 +37,9 @@ class ConcatenateOp(BaseFlatIROp):
         if len(operands) == 1:
             return [operands[0]]
 
-        # TODO: if we could use numpy here, then we could implement the constant folding.
+        # TODO https://github.com/NVIDIA/TensorRT-Incubator/issues/70: if we could use numpy here, then we could implement the constant folding.
         # Otherwise, implement a fold method in MLIR-TRT.
 
-        # Build 'stablehlo.concatenate'.
         output = stablehlo.concatenate(operands, dimension=concatenate_dim)
         # overwrite output type if its shape is inferred
         if self.outputs[0].shape is not None:
